refactor: modularize proxy route
